// frontend/src/client/types.gen.ts
export type NoteItem = {
    title: string;
    content: string;
    flag: number;
    image?: (string | null);
};